Refactor: 코드 리뷰 반영
- SeletBox 폴더 이름 변경
- 배경 파일 Get 에서 가져온 url 사용으로 변경
- error 메세지 state 방식 변경
- 반응형 styled 수정
