+++ src/pages/PaperPostPage/components/CheckBox/CheckBox.jsx
@@ -1,14 +1,21 @@
-import styles from "../SeletBox/SelectBox.module.scss";
+import styles from "../SeletctBox/SelectBox.module.scss";
 import checkImage from "../../../../assets/images/Enabled.svg";
 
 const Checkbox = ({ id, type, color, image, isChecked, onCheckboxChange }) => {
+  const ColorStyle = {
+    background: `${color}`,
+  };
+
+  const imageStyle = {
+    backgroundImage: image ? `url(${image})` : "none",
+  };
+
   return (
     <>
       <label
         htmlFor={id}
-        className={`${styles.selectLabel} ${
-          type === "color" ? styles[color] : styles[image]
-        } ${isChecked ? "checked" : ""}`}
+        className={`${styles.selectLabel} ${isChecked ? "checked" : ""}`}
+        style={type === "color" ? ColorStyle : imageStyle}
       >
         {isChecked && (
           <img className={styles.check} src={checkImage} alt="checked" />
