+++ src/pages/PaperPostPage/components/PaperPostForm.jsx
@@ -1,14 +1,13 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-
 import { postPaper } from "../../../services/api";
 import styles from "./PaperPostForm.module.scss";
-import SelectBox from "./SeletBox/SelectBox";
+import SelectBox from "./SeletctBox/SelectBox";
 import PrimaryButton from "../../../components/UI/PrimaryButton";
 
 function PaperPostForm() {
   const [inputValue, setInputValue] = useState("");
-  const [error, setError] = useState("none");
+  const [error, setError] = useState(false);
   const [isDisabled, setIsDisabled] = useState(true);
   const [backgroundSelection, setBackgroundSelection] = useState({});
   const navigate = useNavigate();
@@ -16,7 +15,7 @@ function PaperPostForm() {
   const handleInputChange = (e) => {
     const value = e.target.value;
     setInputValue(value);
-    setError(value.trim() === "" ? "block" : "none");
+    setError(value.trim() === "" ? true : false);
     setIsDisabled(value.trim() === "");
   };
 
@@ -59,9 +58,9 @@ function PaperPostForm() {
             onChange={handleInputChange}
             className={styles.sendToInput}
           />
-          <div style={{ display: `${error}` }} className={styles.errorMessage}>
-            값을 입력해주세요.
-          </div>
+          {error && (
+            <div className={styles.errorMessage}>값을 입력해주세요.</div>
+          )}
         </div>
         <div className={styles.textBox}>
           <h2 className={styles.title}>배경화면을 선택해 주세요.</h2>
