+++ src/pages/PaperPostPage/components/PaperPostForm.module.scss
@@ -12,7 +12,7 @@
     cursor: not-allowed;
   }
 
-  @media (max-width: 767px) {
+  @media (max-width: 360px) {
     width: calc(100% - 4px);
     margin: 0 2px;
   }
